move shared Message types to own file
