+++ src/components/Message/index.ts
@@ -9,4 +9,5 @@ export * from './MessageRepliesCountButton';
 export * from './MessageSimple';
 export { default as MessageTeam } from './MessageTeam';
 export * from './MessageText';
+export * from './types';
 export * from './utils';
